fix subgraph queries
