+++ packages/deployments/src/types.ts
@@ -153,3 +153,13 @@ export interface Annotation {
   /** The OpenAPI schema that describes the API endpoint at uri. */
   schema: string
 }
+
+export interface RolesModifier {
+  address: `0x${string}`
+  owner: `0x${string}`
+  avatar: `0x${string}`
+  target: `0x${string}`
+  roles: Role[]
+  allowances: Allowance[]
+  multiSendAddresses: `0x${string}`[]
+}
